Refactor: Moved some of the page components to the new 'home' domain.
